requirements.txt
# HTTP client for API requests
requests>=2.28.0

# Environment variable management
python-dotenv>=0.19.0

# Cryptographic operations (EC keys, SPKI) - required for newget.py
cryptography>=3.4.8

# ASN.1 encoding for TNAuthList structures - required for newget.py
pyasn1>=0.4.8
